+++ src/components/LoanCards/HoverLoanCard/InfoPanels/TrusteeInfoPanel.vue
@@ -24,14 +24,8 @@
 						{{ trusteeName }}
 					</span>
 				</div>
-				<div v-if="this.trusteeType">
-					<dt>Trustee type:</dt>
-					<dd class="tw-text-brand tw-my-0.5 tw-capitalize">
-						{{ trusteeType }}
-					</dd>
-				</div>
 				<div v-if="this.trusteeLocation">
-					<dt>Locatiion:</dt>
+					<dt>Location:</dt>
 					<dd class="tw-text-brand tw-my-0.5">
 						{{ trusteeLocation }}
 					</dd>
@@ -132,7 +126,6 @@ export default {
 		return {
 			// borrowerName: '',
 			trusteeName: '',
-			trusteeType: '',
 			trusteeLocation: '',
 			timeOnKiva: '',
 			numBorrowers: '',
@@ -156,7 +149,6 @@ export default {
 		result({ data }) {
 			// this.borrowerName = _get(data, 'lend.loan.name');
 			this.trusteeName = _get(data, 'lend.loan.trustee.organizationName');
-			this.trusteeType = _get(data, 'lend.loan.trustee.trusteeType');
 			this.trusteeLocation = _get(data, 'lend.loan.trustee.trusteeLocation');
 			this.timeOnKiva = _get(data, 'lend.loan.trustee.memberSince');
 			this.numBorrowers = _get(data, 'lend.loan.trustee.trusteeStats.numLoansEndorsedPublic');