+++ src/pages/Lend/LoanChannelCategoryPage.vue
@@ -6,6 +6,7 @@
 		<loan-channel-category-control
 			:enable-loan-tags="enableLoanTags"
 			:enable-loan-card-exp="enableLoanCardExp"
+			:enable-filter-pills="enableFilterPillsTest"
 		/>
 
 		<add-to-basket-interstitial />
@@ -22,6 +23,7 @@ import WwwPage from '@/components/WwwFrame/WwwPage';
 import AddToBasketInterstitial from '@/components/Lightboxes/AddToBasketInterstitial';
 
 import LoanChannelCategoryControl from '@/pages/Lend/LoanChannelCategoryControl';
+import retryAfterExpiredBasket from '@/plugins/retry-after-expired-basket-mixin';
 
 const pageQuery = gql`
 	query LoanChannelCategoryPageExperiments {
@@ -31,6 +33,10 @@ const pageQuery = gql`
 				value
 			}
 			newLoanCard: uiExperimentSetting(key: "new_loan_card") {
+				key
+				value
+			}
+			filterPills: uiExperimentSetting(key: "filter_pills") {
 				key
 				value
 			}
@@ -45,6 +51,7 @@ export default {
 		LoanChannelCategoryControl,
 		WwwPage,
 	},
+	mixins: [retryAfterExpiredBasket],
 	inject: ['apollo', 'cookieStore'],
 	data() {
 		return {
@@ -55,6 +62,7 @@ export default {
 			pageLayout: 'control',
 			enableLoanTags: false,
 			enableLoanCardExp: false,
+			enableFilterPillsTest: false,
 		};
 	},
 	apollo: {
@@ -65,6 +73,7 @@ export default {
 				return Promise.all([
 					client.query({ query: experimentAssignmentQuery, variables: { id: 'loan_tags' } }),
 					client.query({ query: experimentAssignmentQuery, variables: { id: 'new_loan_card' } }),
+					client.query({ query: experimentAssignmentQuery, variables: { id: 'filter_pills' } }),
 				]);
 			});
 		}
@@ -82,6 +91,9 @@ export default {
 
 		// Initialize New Loan Card Experiment
 		this.initializeNewLoanCardTest();
+
+		// Initialize Filter Pills Experimentx
+		this.initializeFilterPillsTest();
 	},
 	methods: {
 		initializeNewLoanCardTest() {
@@ -109,6 +121,20 @@ export default {
 					'Lending',
 					'EXP-CORE-792-Oct2022',
 					loanTagsExperiment.version
+				);
+			}
+		},
+		initializeFilterPillsTest() {
+			const filterPilssExp = this.apollo.readFragment({
+				id: 'Experiment:filter_pills',
+				fragment: experimentVersionFragment,
+			}) || {};
+			this.enableFilterPillsTest = filterPilssExp.version === 'b';
+			if (filterPilssExp.version) {
+				this.$kvTrackEvent(
+					'Lending',
+					'EXP-CORE-1195-Mar2023',
+					filterPilssExp.version
 				);
 			}
 		},