+++ src/pages/Portfolio/PortfolioOverviewPage.vue
@@ -1,0 +1,45 @@
+<template>
+	<www-page main-class="tw-bg-secondary">
+		<template #secondary>
+			<the-my-kiva-secondary-menu />
+		</template>
+		<kv-page-container>
+			<kv-grid class="tw-grid-cols-12 tw--mx-2.5 md:tw-mx-0">
+				<the-portfolio-tertiary-menu class="tw-pt-2 tw-col-span-3 tw-hidden md:tw-block" />
+				<div class="tw-col-span-12 md:tw-col-span-9 tw-pt-3">
+					<account-overview />
+					<recent-loans-list />
+					<!-- rec'd loans -->
+					<!-- kiva effect -->
+					<!-- updates -->
+					<!-- lending insights -->
+					<!-- teams -->
+					<!-- stat graphs -->
+				</div>
+			</kv-grid>
+		</kv-page-container>
+	</www-page>
+</template>
+
+<script>
+import WwwPage from '@/components/WwwFrame/WwwPage';
+import TheMyKivaSecondaryMenu from '@/components/WwwFrame/Menus/TheMyKivaSecondaryMenu';
+import ThePortfolioTertiaryMenu from '@/components/WwwFrame/Menus/ThePortfolioTertiaryMenu';
+import KvGrid from '~/@kiva/kv-components/vue/KvGrid';
+import KvPageContainer from '~/@kiva/kv-components/vue/KvPageContainer';
+import AccountOverview from './AccountOverview';
+import RecentLoansList from './RecentLoansList';
+
+export default {
+	name: 'PortfolioOverviewPage',
+	components: {
+		AccountOverview,
+		KvGrid,
+		KvPageContainer,
+		RecentLoansList,
+		TheMyKivaSecondaryMenu,
+		ThePortfolioTertiaryMenu,
+		WwwPage,
+	},
+};
+</script>